populate

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -28,14 +29,21 @@ export const postUpload = async (req, res) => {
       title, description, hashtags
     },
     file,
+    session: {
+      user: { _id },
+    }
   } = req;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       hashtags: Video.formatHashtags(hashtags),
       fileUrl: file.path,
+      owner: _id,
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch(error) {
     console.log(`postUpdate error: ${error}`);
@@ -44,10 +52,17 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
-  const video = Video.exists({_id: id});
+  const {
+    params: { id },
+    session: { user: { _id } },
+  } = req;
+  const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video Not Found" });
+  }
+  console.log(video.owner, _id);
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).render("404", { pageTitle: "You are not the owner of video."});
   }
   await Video.findByIdAndRemove(id);
   return res.redirect("/");
@@ -55,7 +70,7 @@ export const deleteVideo = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video Not Found" });
   }
@@ -63,20 +78,32 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: { user: { _id } },
+  } = req;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video Not Found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(400).render("404", { pageTitle: "You are not the owner of video."});
   }
   return res.render("videos/edit", { pageTitle: "Edit video", video })
 };
 
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
-  const video = Video.exists({_id: id});
+  const {
+    params: { id },
+    body: { title, description, hashtags },
+    session: { user: { _id } },
+  } = req;
+  const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video Not Found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(400).render("404", { pageTitle: "You are not the owner of video."});
   }
   await Video.findByIdAndUpdate(id, {
     title, description,
